+++ test/run.js
@@ -1,5 +1,5 @@
-const extOs = require('../lib/os.js');
-const path = require('path');
+const extOs = require('../lib/os.js')
+const path = require('path')
 
-const iPkg = path.join(__dirname, './case/package-from-github/package.json');
-extOs.installPackage(iPkg, {production: true});
+const iPkg = path.join(__dirname, './case/package-from-github/package.json')
+extOs.installPackage(iPkg, { production: true })